headings all use the same method now
